requirements.txt
Django==2.1.2
django_debug_toolbar==1.10.1
numpy==1.11.1
